[manhattan] Organize package
For consistency with other packages, define components in
their own files and have index.js only contain exports.
